+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow setting custom AMQP message expiry with env var RPC_AMQP_MESSAGE_EXPIRY
 
 ## [1.0.1] - 2021-12-16